+++ src/app/manage/clients/[id]/page.tsx
